feat: rebuild loupedeck plugin as single implementation
