+++ satellite/package.json
@@ -42,7 +42,7 @@
 		"@julusian/image-rs": "^2.0.2",
 		"@julusian/jpeg-turbo": "^2.3.0",
 		"@julusian/segfault-raub": "^2.3.2",
-		"@loupedeck/node": "^1.2.0",
+		"@loupedeck/node": "^2.0.3",
 		"@napi-rs/canvas": "^0.1.80",
 		"@xencelabs-quick-keys/node": "^1.0.0",
 		"conf": "^15.0.0",
