+++ satellite/src/device-types/loupedeck-plugin.ts
@@ -1,7 +1,8 @@
 import {
 	listLoupedecks,
+	LoupedeckBufferFormat,
 	LoupedeckDevice,
-	LoupedeckModelId,
+	LoupedeckDisplayId,
 	openLoupedeck,
 	type LoupedeckDeviceInfo,
 } from '@loupedeck/node'
@@ -12,11 +13,13 @@ import type {
 	DeviceRegisterProps,
 	OpenSurfaceResult,
 	SurfaceContext,
+	DeviceDrawProps,
 } from './api.js'
-import { compileLoupedeckLiveProps, LoupedeckLiveWrapper } from './loupedeck-live.js'
 import { assertNever } from '../lib.js'
-import { compileLoupedeckLiveSProps, LoupedeckLiveSWrapper } from './loupedeck-live-s.js'
-import { compileRazerStreamControllerXProps, RazerStreamControllerXWrapper } from './razer-stream-controller-x.js'
+import { SatelliteSurfaceLayout } from '../generated/SurfaceManifestSchema.js'
+import { Pincode5x3 } from './pincode.js'
+import { CardGenerator } from '../graphics/cards.js'
+import { parseColor } from './lib.js'
 
 export const LOUPEDECK_PLUGIN_ID = 'loupedeck'
 
@@ -53,35 +56,198 @@ export class LoupedeckPlugin implements SurfacePlugin<LoupedeckDeviceInfo> {
 		pluginInfo: LoupedeckDeviceInfo,
 		context: SurfaceContext,
 	): Promise<OpenSurfaceResult> => {
-		let factory: new (deviceId: string, device: LoupedeckDevice, context: SurfaceContext) => SurfaceInstance
-		let propsFactory: (device: LoupedeckDevice) => DeviceRegisterProps
-
-		switch (pluginInfo.model) {
-			case LoupedeckModelId.LoupedeckLive:
-			case LoupedeckModelId.RazerStreamController:
-				factory = LoupedeckLiveWrapper
-				propsFactory = compileLoupedeckLiveProps
-				break
-			case LoupedeckModelId.LoupedeckLiveS:
-				factory = LoupedeckLiveSWrapper
-				propsFactory = compileLoupedeckLiveSProps
-				break
-			case LoupedeckModelId.RazerStreamControllerX:
-				factory = RazerStreamControllerXWrapper
-				propsFactory = compileRazerStreamControllerXProps
-				break
-			case LoupedeckModelId.LoupedeckCt:
-			case LoupedeckModelId.LoupedeckCtV1:
-				throw new Error('Unsupported model')
-			default:
-				assertNever(pluginInfo.model)
-				throw new Error('Unsupported model')
-		}
-
 		const loupedeck = await openLoupedeck(pluginInfo.path)
 		return {
-			surface: new factory(surfaceId, loupedeck, context),
-			registerProps: propsFactory(loupedeck),
+			surface: new LoupedeckWrapper(surfaceId, loupedeck, context),
+			registerProps: compileLoupedeckProps(loupedeck),
 		}
+	}
+}
+
+function compileLoupedeckProps(device: LoupedeckDevice): DeviceRegisterProps {
+	const surfaceManifest: SatelliteSurfaceLayout = {
+		stylePresets: {
+			default: {
+				bitmap: {
+					w: device.lcdKeySize,
+					h: device.lcdKeySize,
+				},
+			},
+			button: {
+				colors: 'hex',
+			},
+			empty: {},
+		},
+		controls: {},
+	}
+
+	for (const control of device.controls) {
+		const { row, column } = control
+
+		switch (control.type) {
+			case 'button':
+				surfaceManifest.controls[control.id] = {
+					row,
+					column,
+					stylePreset: control.feedbackType === 'rgb' ? 'button' : 'default',
+				}
+				break
+			case 'encoder':
+				surfaceManifest.controls[control.id] = { row, column, stylePreset: 'empty' }
+				break
+			case 'wheel':
+				if (device.displayWheel) {
+					surfaceManifest.stylePresets.wheel = {
+						bitmap: {
+							w: device.displayWheel.width,
+							h: device.displayWheel.height,
+						},
+					}
+					surfaceManifest.controls[control.id] = { row, column, stylePreset: 'wheel' }
+				}
+				break
+			default:
+				assertNever(control)
+				break
+		}
+	}
+
+	return {
+		brightness: true,
+		surfaceManifest,
+		pincodeMap: Pincode5x3(1),
+	}
+}
+
+class LoupedeckWrapper implements SurfaceInstance {
+	readonly pluginId = LOUPEDECK_PLUGIN_ID
+
+	readonly #deck: LoupedeckDevice
+	readonly #surfaceId: string
+
+	public get surfaceId(): string {
+		return this.#surfaceId
+	}
+	public get productName(): string {
+		return this.#deck.modelName
+	}
+
+	public constructor(surfaceId: string, device: LoupedeckDevice, context: SurfaceContext) {
+		this.#deck = device
+		this.#surfaceId = surfaceId
+
+		this.#deck.on('error', (e) => context.disconnect(e))
+
+		this.#deck.on('down', (control) => {
+			context.keyDownById(control.id)
+		})
+		this.#deck.on('up', (control) => {
+			context.keyUpById(control.id)
+		})
+		this.#deck.on('rotate', (control, delta) => {
+			if (delta < 0) {
+				context.rotateLeftById(control.id)
+			} else if (delta > 0) {
+				context.rotateRightById(control.id)
+			}
+		})
+		this.#deck.on('touchstart', (data) => {
+			for (const touch of data.changedTouches) {
+				if (touch.target.control !== undefined) {
+					context.keyDownById(touch.target.control.id)
+				} else if (touch.target.screen == LoupedeckDisplayId.Wheel) {
+					const wheelControl = this.#deck.controls.find((c) => c.type === 'wheel')
+					if (wheelControl) context.keyDownById(wheelControl.id)
+				}
+			}
+		})
+		this.#deck.on('touchend', (data) => {
+			for (const touch of data.changedTouches) {
+				if (touch.target.control !== undefined) {
+					context.keyUpById(touch.target.control.id)
+				} else if (touch.target.screen == LoupedeckDisplayId.Wheel) {
+					const wheelControl = this.#deck.controls.find((c) => c.type === 'wheel')
+					if (wheelControl) context.keyUpById(wheelControl.id)
+				}
+			}
+		})
+	}
+
+	async close(): Promise<void> {
+		await this.#deck.blankDevice(true, true).catch(() => null)
+
+		await this.#deck.close()
+	}
+	async initDevice(): Promise<void> {
+		// Start with blanking it
+		await this.blankDevice()
+	}
+
+	async deviceAdded(): Promise<void> {}
+	async setBrightness(percent: number): Promise<void> {
+		await this.#deck.setBrightness(percent / 100)
+	}
+	async blankDevice(skipButtons?: boolean): Promise<void> {
+		await this.#deck.blankDevice(true, !skipButtons)
+	}
+	async draw(_signal: AbortSignal, d: DeviceDrawProps): Promise<void> {
+		const control = this.#deck.controls.find((c) => c.id === d.controlId)
+		if (!control) return
+
+		if (control.type === 'button') {
+			if (control.feedbackType === 'rgb') {
+				const color = parseColor(d.color)
+
+				await this.#deck.setButtonColor({
+					id: d.controlId,
+					red: color.r,
+					green: color.g,
+					blue: color.b,
+				})
+			} else if (control.feedbackType === 'lcd') {
+				if (d.image) {
+					const buffer = await d.image(this.#deck.lcdKeySize, this.#deck.lcdKeySize, 'rgb')
+					await this.#deck.drawKeyBuffer(d.controlId, buffer, LoupedeckBufferFormat.RGB)
+				} else {
+					throw new Error(`Cannot draw for Loupedeck without image`)
+				}
+			}
+		} else if (control.type === 'wheel') {
+			if (!this.#deck.displayWheel) return
+
+			const width = this.#deck.displayWheel.width
+			const height = this.#deck.displayWheel.height
+
+			if (d.image) {
+				const buffer = await d.image(width, height, 'rgb')
+				await this.#deck.drawBuffer(
+					LoupedeckDisplayId.Wheel,
+					buffer,
+					LoupedeckBufferFormat.RGB,
+					width,
+					height,
+					0,
+					0,
+				)
+			} else {
+				throw new Error(`Cannot draw for Loupedeck without image`)
+			}
+		}
+	}
+
+	async showStatus(
+		signal: AbortSignal,
+		cardGenerator: CardGenerator,
+		hostname: string,
+		status: string,
+	): Promise<void> {
+		const width = this.#deck.displayMain.width
+		const height = this.#deck.displayMain.height
+
+		const buffer = await cardGenerator.generateBasicCard(width, height, 'rgb', hostname, status)
+
+		if (signal.aborted) return
+
+		await this.#deck.drawBuffer(LoupedeckDisplayId.Center, buffer, LoupedeckBufferFormat.RGB, width, height, 0, 0)
 	}
 }
